Extract restricted editing commands tests.
